Remove splitting of Bytes
Turns out this is unneeded when you don't try to outsmart the elm
compiler

I tried writing a custom `iterate` that minimized allocation. It looked
all tail-recursive, but in fact wasn't. That could cause a
RangeException which would lead to a (silent) decoding failure.

By instead using Decode.loop all is good, and this complexity is not
needed any more.